backend/requirements_3_12.txt
alembic==1.14.0
bcrypt==4.2.1
blinker==1.9.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
commonmark==0.9.1
Deprecated==1.2.15
dotenv==0.9.9
Flask==3.1.0
Flask-Bcrypt==1.0.1
Flask-Cors==5.0.0
Flask-JWT-Extended==4.7.1
Flask-Limiter==3.9.2
flask-marshmallow==1.2.1
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
idna==3.10
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
limits==3.14.1
Mako==1.3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.23.1
mdurl==0.1.2
ordered-set==4.1.0
packaging==24.2
pluggy==1.5.0
psycopg2==2.9.10
Pygments==2.18.0
PyJWT==2.10.1
pytest==8.3.4
pytest-flask==1.3.0
python-dotenv==1.0.1
requests==2.32.3
rich==13.9.4
six==1.17.0
SQLAlchemy==2.0.36
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.1.3
wrapt==1.17.0
